core/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile core/requirements.in -o core/requirements.txt
alembic==1.13.2
    # via -r core/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asyncpg==0.29.0
    # via -r core/requirements.in
bcrypt==4.1.2
    # via pwdlib
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==43.0.1
    # via pyjwt
dnspython==2.6.1
    # via email-validator
email-validator==2.1.2
    # via
    #   fastapi
    #   fastapi-users
fastapi==0.112.3
    # via
    #   -r core/requirements.in
    #   fastapi-users
fastapi-cli==0.0.5
    # via fastapi
fastapi-users==13.0.0
    # via
    #   -r core/requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==6.0.1
    # via fastapi-users
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.8
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via fastapi
makefun==1.15.4
    # via fastapi-users
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.7
    # via fastapi
pwdlib==0.2.0
    # via fastapi-users
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-extra-types==2.9.0
    # via fastapi
pydantic-settings==2.4.0
    # via fastapi
pygments==2.18.0
    # via rich
pyjwt==2.8.0
    # via fastapi-users
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via
    #   fastapi
    #   fastapi-users
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
rich==13.8.0
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.34
    # via
    #   -r core/requirements.in
    #   alembic
    #   fastapi-users-db-sqlalchemy
starlette==0.38.4
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
ujson==5.10.0
    # via fastapi
uvicorn==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
